- Made all (relevant) code compile at Visual Studio 2012's warning level 4. - Changed type of st_ucon64_t::parport from int to uint16_t.

+++ src/backup/doctor64.h
@@ -24,8 +24,8 @@ Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 extern const st_getopt2_t doctor64_usage[];
 
 #ifdef  USE_PARALLEL
-extern int doctor64_read (const char *filename, unsigned int parport);
-extern int doctor64_write (const char *filename, int start, int len, unsigned int parport);
+extern int doctor64_read (const char *filename, unsigned short parport);
+extern int doctor64_write (const char *filename, int start, int len, unsigned short parport);
 #endif
 
 #endif
